add create_subnetwork as RiverNetwork method
