+++ src/earthkit/hydro/__init__.py
@@ -1,5 +1,4 @@
 from .readers import from_cama_downxy, from_cama_nextxy, from_d8, load_river_network, create_river_network
 from .river_network import RiverNetwork
 from .accumulation import accuflux, upstream, downstream
-from .subnetwork import create_subnetwork
 from .catchment import catchment, subcatchment
